requirements.txt
datettime==5.5
h5py==3.12.1
matplotlib==3.9.2
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
scikit-learn==1.5.2
tqdm==4.67.0
torch
# # --index-url https://download.pytorch.org/whl/cu117 based on the cuda requirements
# git+https://github.com/hassony2/torch_videovision
# h5py
# matplotlib
# numpy
# opencv-python
# pandas
# scikit-learn
# seaborn
# tensorboard
# torch
# torchaudio
# torchvision
# torchsummary
# tqdm